r5x: Add QTI health implementation
Add QTI health implementation so that it waits for battery power
supply for a max timeout of 5 seconds. This is to ensure by the
time health HAL service starts, battery power supply device is
available. Without this, health HAL couldn't find battery power
supply sometimes because of the time taken for battery charger
driver to probe on platforms where it can happen only after the
remote subsytem is bought out of reset.